+++ src/inventory/mod.rs
@@ -14,6 +14,6 @@
 /// The default location to inventory.
 pub const DEFAULT_INVENTORY_PATH: &str = "inventory.yaml";
 /// The default interval for record checking.
-pub const DEFAULT_WATCH_INTERVAL: u32 = 5000;
+pub const DEFAULT_WATCH_INTERVAL: u64 = 5000;
 
 pub mod models;
